app/requirements-dev.txt
-r requirements.txt
pytest>=7.0
pytest-django>=4.5
